@@ -1,3 +1,3 @@
 module linkedlist
 
-go 1.16
+go 1.18
